backend/requirements.txt
pymupdf==1.23.26
pytesseract==0.3.10
Pillow==10.2.0
